requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow==2.18.0
keras>=3.0.0
Pillow>=10.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
streamlit>=1.33.0
